+++ classes/class.ilH5PPageComponentPlugin.php
@@ -2,14 +2,25 @@
 
 declare(strict_types=1);
 
+require_once __DIR__ . "/../vendor/autoload.php";
+
 use srag\Plugins\H5P\CI\Rector\DICTrait\Replacement\VersionComparator;
 use srag\Plugins\H5P\Content\IContentRepository;
 use srag\Plugins\H5P\Content\IContent;
 use srag\Plugins\H5P\ArrayBasedRequestWrapper;
 use srag\Plugins\H5P\RequestHelper;
 use Psr\Http\Message\ServerRequestInterface;
+use ILIAS\DI\Container;
 
 /**
+ * This plugin can only be installed if the H5P main plugin is available on the
+ * file-system.
+ *
+ * However, the plugin may be installed before the main plugin is installed, in
+ * which case we need to disable the plugins functionality to prevent fatals.
+ * This is why in this class and the GUI class we need to check if the main has
+ * been installed.
+ *
  * @author       Thibeau Fuhrer <thibeau@sr.solutions>
  * @noinspection AutoloadingIssuesInspection
  */
@@ -19,6 +30,8 @@ class ilH5PPageComponentPlugin extends ilPageComponentPlugin
     public const PLUGIN_ID = "pchfp";
 
     public const PROPERTY_CONTENT_ID = 'content_id';
+
+    protected const H5P_MAIN_AUTOLOAD = __DIR__ . "/../../../../Repository/RepositoryObject/H5P/vendor/autoload.php";
 
     /**
      * @var self|null
@@ -51,19 +64,19 @@ class ilH5PPageComponentPlugin extends ilPageComponentPlugin
     protected $ctrl;
 
     /**
-     * @throws LogicException if the main plugin is not installed.
+     * @throws LogicException if the main plugin is not available (not found in file-system).
      */
     public function __construct()
     {
         global $DIC;
         parent::__construct();
 
-        if (!class_exists('ilH5PPlugin')) {
+        if (!file_exists(self::H5P_MAIN_AUTOLOAD)) {
             throw new LogicException("You cannot use this plugin without installing the main plugin first.");
         }
 
-        if (!$this->isActive()) {
-            return;
+        if (!$this->isMainPluginLoaded()) {
+            require_once self::H5P_MAIN_AUTOLOAD;
         }
 
         $container = ilH5PPlugin::getInstance()->getContainer();
@@ -109,10 +122,30 @@ class ilH5PPageComponentPlugin extends ilPageComponentPlugin
     }
 
     /**
+     * Exchanges the default renderer if the main plugin is available but not active.
+     * Otherwise, this plugin cannot render any H5P contents and could even lead to fatals
+     * because we are using custom UI components, which arent renderable by default.
+     *
+     * @inheritDoc
+     */
+    public function exchangeUIRendererAfterInitialization(Container $dic): Closure
+    {
+        if ($this->isMainPluginLoaded() && !$this->isMainPluginActive()) {
+            return ilH5PPlugin::getInstance()->exchangeUIRendererAfterInitialization($dic);
+        }
+
+        return parent::exchangeUIRendererAfterInitialization($dic);
+    }
+
+    /**
      * @inheritDoc
      */
     public function onClone(&$a_properties, $a_plugin_version): void
     {
+        if (!$this->isMainPluginInstalled()) {
+            return;
+        }
+
         $content_id = (int) $a_properties[self::PROPERTY_CONTENT_ID];
 
         $content = $this->content_repository->getContent($content_id);
@@ -142,6 +175,10 @@ class ilH5PPageComponentPlugin extends ilPageComponentPlugin
      */
     public function onDelete($a_properties, $a_plugin_version): void
     {
+        if (!$this->isMainPluginInstalled()) {
+            return;
+        }
+
         $content_id = (int) $a_properties[self::PROPERTY_CONTENT_ID];
 
         if ($this->shouldDeleteContent()) {
@@ -197,5 +234,20 @@ class ilH5PPageComponentPlugin extends ilPageComponentPlugin
     protected function isContentIdCut(int $content_id): bool
     {
         return (true === ilSession::get(self::PLUGIN_ID . '_cut_old_content_id_' . $content_id));
+    }
+
+    private function isMainPluginLoaded(): bool
+    {
+        return class_exists('ilH5PPlugin');
+    }
+
+    private function isMainPluginInstalled(): bool
+    {
+        return ilH5PPlugin::getInstance()->getContainer()->getRepositoryFactory()->general()->isMainPluginInstalled();
+    }
+
+    private function isMainPluginActive(): bool
+    {
+        return ilH5PPlugin::getInstance()->isActive();
     }
 }